organize reducer & add save of theme settings
