+++ libs/redux/saveLocalStorage.ts
@@ -1,0 +1,5 @@
+"use client";
+
+export function saveThemeLS(theme: boolean) {
+  localStorage.setItem("theme", JSON.stringify(theme));
+}
